+++ CHANGELOG.md
@@ -17,3 +17,9 @@ All notable changes to `dependencies` will be documented in this file
 ## 0.2.0 - 2021-07-09
 - refactor test suite to use `Http` facade instead of `GuzzleHttp\Client`
 - refactor `DependencyService` methods to return an instance of `DependencyUrl` for retrieving urls & svgs
+
+
+## 0.2.1 - 2021-07-12
+- make `DependencyUrlTest` to test url generators in isolation
+- make `DependencySvgTest` to test svg generators in isolation
+- make `ConfigTest`